Let taglist use Homebrew version of ctags.
The version of ctags in standard MacOSX distribution does not
suffice the need of the vim plugin "taglist."

To install the latest version of ctags, Homebrew is used.

Here we specify the path of ctags used by the taglist plugin.
Note we do not change the ordering of elements in $PATH,
because different tools may require different ordering.
That will be hard to maintain.